python-agents/ingestion-agent/requirements.txt
# Pinned for a clean resolve without optional heavy deps (e.g. pyiceberg on newer supabase).
httpx==0.25.2
supabase==2.4.0
python-dotenv==1.0.0
schedule==1.2.1
